Server/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
psycopg2==2.9.9
requests==2.31.0
httpx==0.27.0
python-dotenv==1.0.1